Update Playlist.js
Step 34 created a conflict between TrackList being called in SearchResults as well as Playlist.
By commenting out Tracklist in Playlist the page now displays properly.
I'm assuming that later steps will resolve the problem that step 34 caused with the display, but for now commenting out the TrackList component in Playlist is the working solution.